pyproject.toml
[project]
name = "ragpipeline"
version = "0.1.0"
description = ""
authors = [
    {name = "Marv-11",email = "shatha.alkhudhuri@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.135.2,<0.136.0)",
    "uvicorn (>=0.42.0,<0.43.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "python-multipart (>=0.0.9,<0.0.10)",
    "aiofiles (>=25.1.0,<26.0.0)",
    "langchain (>=1.2.13,<2.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "langchain-text-splitters (>=1.1.1,<2.0.0)",
    "pypdf (>=6.9.2,<7.0.0)",
    "motor (>=3.7.1,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = ">=3.12,<4.0.0"
